pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "goose-calendar-extension"
version = "0.1.0"
description = "Goose extension for Google Calendar management"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "goose-ai",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "python-dateutil",
    "mcp[cli]>=1.2.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/goose-calendar-extension"
"Bug Reports" = "https://github.com/yourusername/goose-calendar-extension/issues"
"Source" = "https://github.com/yourusername/goose-calendar-extension"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[project.scripts]
goose-calendar = "goose_calendar:main"
